Leistungsberechnung, Fahrergweicht

+++ src/main/java/org/pirate/sportwebserver/dto/strava/ActivityStreams.java
@@ -2,7 +2,8 @@ package org.pirate.sportwebserver.dto.strava;
 
 import java.util.List;
 
-public class ActivityStreams {
+public class ActivityStreams
+{
 
 	private StreamData<List<Double>> latlng;
 
@@ -14,63 +15,77 @@ public class ActivityStreams {
 
 	private StreamData<Integer> heartrate;
 
-	private StreamData<Integer> watts;
+	private StreamData<Double> watts;
 
 	private StreamData<Double> velocity_smooth;
 
-	public StreamData<List<Double>> getLatlng() {
+	public StreamData<List<Double>> getLatlng()
+	{
 		return latlng;
 	}
 
-	public void setLatlng(StreamData<List<Double>> latlng) {
+	public void setLatlng(StreamData<List<Double>> latlng)
+	{
 		this.latlng = latlng;
 	}
 
-	public StreamData<Double> getDistance() {
+	public StreamData<Double> getDistance()
+	{
 		return distance;
 	}
 
-	public void setDistance(StreamData<Double> distance) {
+	public void setDistance(StreamData<Double> distance)
+	{
 		this.distance = distance;
 	}
 
-	public StreamData<Integer> getTime() {
+	public StreamData<Integer> getTime()
+	{
 		return time;
 	}
 
-	public void setTime(StreamData<Integer> time) {
+	public void setTime(StreamData<Integer> time)
+	{
 		this.time = time;
 	}
 
-	public StreamData<Double> getAltitude() {
+	public StreamData<Double> getAltitude()
+	{
 		return altitude;
 	}
 
-	public void setAltitude(StreamData<Double> altitude) {
+	public void setAltitude(StreamData<Double> altitude)
+	{
 		this.altitude = altitude;
 	}
 
-	public StreamData<Integer> getHeartrate() {
+	public StreamData<Integer> getHeartrate()
+	{
 		return heartrate;
 	}
 
-	public void setHeartrate(StreamData<Integer> heartrate) {
+	public void setHeartrate(StreamData<Integer> heartrate)
+	{
 		this.heartrate = heartrate;
 	}
 
-	public StreamData<Integer> getWatts() {
+	public StreamData<Double> getWatts()
+	{
 		return watts;
 	}
 
-	public void setWatts(StreamData<Integer> watts) {
+	public void setWatts(StreamData<Double> watts)
+	{
 		this.watts = watts;
 	}
 
-	public StreamData<Double> getVelocity_smooth() {
+	public StreamData<Double> getVelocity_smooth()
+	{
 		return velocity_smooth;
 	}
 
-	public void setVelocity_smooth(StreamData<Double> velocity_smooth) {
+	public void setVelocity_smooth(StreamData<Double> velocity_smooth)
+	{
 		this.velocity_smooth = velocity_smooth;
 	}
 }
